TW-898: move logic of download+preview file from chat controller to mixin
