+++ lib/pages/chat/chat.dart
@@ -8,14 +8,6 @@ import 'package:desktop_drop/desktop_drop.dart';
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
 import 'package:fluffychat/di/global/get_it_initializer.dart';
-import 'package:fluffychat/domain/app_state/preview_file/download_file_for_preview_failure.dart';
-import 'package:fluffychat/domain/app_state/preview_file/download_file_for_preview_loading.dart';
-import 'package:fluffychat/domain/app_state/preview_file/download_file_for_preview_success.dart';
-import 'package:fluffychat/domain/model/download_file/download_file_for_preview_response.dart';
-import 'package:fluffychat/domain/model/extensions/mime_type_extension.dart';
-import 'package:fluffychat/domain/model/preview_file/document_uti.dart';
-import 'package:fluffychat/domain/model/preview_file/supported_preview_file_types.dart';
-import 'package:fluffychat/domain/usecase/download_file_for_preview_interactor.dart';
 import 'package:fluffychat/domain/usecase/room/chat_get_pinned_events_interactor.dart';
 import 'package:fluffychat/domain/usecase/send_file_interactor.dart';
 import 'package:fluffychat/pages/chat/chat_context_menu_actions.dart';
@@ -41,8 +33,6 @@ import 'package:fluffychat/utils/matrix_sdk_extensions/event_extension.dart';
 import 'package:fluffychat/utils/matrix_sdk_extensions/ios_badge_client_extension.dart';
 import 'package:fluffychat/utils/matrix_sdk_extensions/matrix_locals.dart';
 import 'package:fluffychat/utils/network_connection_service.dart';
-import 'package:fluffychat/utils/permission_dialog.dart';
-import 'package:fluffychat/utils/permission_service.dart';
 import 'package:fluffychat/utils/platform_infos.dart';
 import 'package:fluffychat/utils/responsive/responsive_utils.dart';
 import 'package:fluffychat/utils/twake_snackbar.dart';
@@ -54,7 +44,6 @@ import 'package:fluffychat/widgets/mxc_image.dart';
 import 'package:fluffychat/widgets/twake_app.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 
 import 'package:go_router/go_router.dart';
@@ -62,14 +51,10 @@ import 'package:image_picker/image_picker.dart';
 import 'package:linagora_design_flutter/images_picker/asset_counter.dart';
 import 'package:linagora_design_flutter/images_picker/images_picker_grid.dart';
 import 'package:matrix/matrix.dart';
-import 'package:open_file/open_file.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:permission_handler/permission_handler.dart';
 import 'package:record/record.dart';
 import 'package:scroll_to_index/scroll_to_index.dart';
-import 'package:share_plus/share_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:universal_html/html.dart' as html;
+import 'package:flutter_gen/gen_l10n/l10n.dart';
 
 import '../../utils/account_bundles.dart';
 import '../../utils/localized_exception_extension.dart';
@@ -470,137 +455,6 @@ class ChatController extends State<Chat>
       editEvent = null;
       pendingText = '';
     });
-  }
-
-  void onFileTapped(Event event) {
-    if (PlatformInfos.isWeb) {
-      onFileTappedWeb(event);
-    } else {
-      onFileTappedMobile(event);
-    }
-  }
-
-  void onFileTappedWeb(Event event) {
-    return _handlePreviewWeb(event: event);
-  }
-
-  void onFileTappedMobile(Event event) async {
-    final permissionHandler = PermissionHandlerService();
-    if (await permissionHandler.noNeedStoragePermission()) {
-      return _handleDownloadFileForPreviewMobile(event: event);
-    }
-    final storagePermissionStatus =
-        await permissionHandler.storagePermissionStatus;
-    switch (storagePermissionStatus) {
-      case PermissionStatus.denied:
-        await showDialog(
-          useRootNavigator: false,
-          context: context,
-          builder: (context) {
-            return PermissionDialog(
-              permission: Permission.storage,
-              explainTextRequestPermission:
-                  Text(L10n.of(context)!.explainStoragePermission),
-              icon: const Icon(Icons.preview_outlined),
-            );
-          },
-        );
-        if (await permissionHandler.storagePermissionStatus ==
-            PermissionStatus.granted) {
-          _handleDownloadFileForPreviewMobile(event: event);
-        }
-        break;
-      case PermissionStatus.granted:
-        _handleDownloadFileForPreviewMobile(event: event);
-        break;
-      case PermissionStatus.permanentlyDenied:
-        showDialog(
-          useRootNavigator: false,
-          context: context,
-          builder: (context) {
-            return PermissionDialog(
-              permission: Permission.storage,
-              explainTextRequestPermission:
-                  Text(L10n.of(context)!.explainGoToStorageSetting),
-              icon: const Icon(Icons.preview_outlined),
-            );
-          },
-        );
-        break;
-      case PermissionStatus.restricted:
-      case PermissionStatus.limited:
-      case PermissionStatus.provisional:
-        break;
-      default:
-        break;
-    }
-  }
-
-  void _handlePreviewWeb({required Event event}) async {
-    if (!event.hasAttachment) {
-      TwakeSnackBar.show(context, L10n.of(context)!.errorPreviewingFile);
-      return;
-    }
-
-    if (event.isPdfFile()) {
-      return previewPdfWeb(context, event);
-    }
-
-    downloadFileAction(context, event);
-  }
-
-  void _handleDownloadFileForPreviewMobile({required Event event}) async {
-    final downloadFileForPreviewInteractor =
-        getIt.get<DownloadFileForPreviewInteractor>();
-    final tempDirPath = (await getTemporaryDirectory()).path;
-    downloadFileForPreviewInteractor
-        .execute(
-      event: event,
-      tempDirPath: tempDirPath,
-    )
-        .listen((event) {
-      event.fold((failure) {
-        if (failure is DownloadFileForPreviewFailure) {
-          TwakeSnackBar.show(context, 'Error: ${failure.exception}');
-        }
-        TwakeDialog.hideLoadingDialog(context);
-      }, (success) {
-        if (success is DownloadFileForPreviewSuccess) {
-          _openDownloadedFileForPreview(
-            downloadFileForPreviewResponse:
-                success.downloadFileForPreviewResponse,
-          );
-          TwakeDialog.hideLoadingDialog(context);
-        } else if (success is DownloadFileForPreviewLoading) {
-          TwakeDialog.showLoadingDialog(context);
-        }
-      });
-    });
-  }
-
-  void _openDownloadedFileForPreview({
-    required DownloadFileForPreviewResponse downloadFileForPreviewResponse,
-  }) async {
-    final mimeType = downloadFileForPreviewResponse.mimeType;
-    if (Platform.isAndroid &&
-        SupportedPreviewFileTypes.apkMimeTypes.contains(mimeType)) {
-      await Share.shareXFiles([XFile(downloadFileForPreviewResponse.filePath)]);
-      return;
-    }
-    final openResults = await OpenFile.open(
-      downloadFileForPreviewResponse.filePath,
-      type: mimeType,
-      uti: DocumentUti(SupportedPreviewFileTypes.iOSSupportedTypes[mimeType])
-          .value,
-    );
-    Logs().d(
-      'ChatController:_openDownloadedFileForPreview(): ${openResults.message}',
-    );
-
-    if (openResults.type != ResultType.done) {
-      await Share.shareXFiles([XFile(downloadFileForPreviewResponse.filePath)]);
-      return;
-    }
   }
 
   void openVideoCameraAction() async {
@@ -1537,20 +1391,6 @@ class ChatController extends State<Chat>
 
   Future<String?> downloadFileAction(BuildContext context, Event event) async =>
       await event.saveFile(context);
-
-  void previewPdfWeb(BuildContext context, Event event) async {
-    final pdf = await event.getFile(context);
-    if (pdf.result == null || event.sizeString != pdf.result?.sizeString) {
-      TwakeSnackBar.show(context, L10n.of(context)!.errorGettingPdf);
-
-      return;
-    }
-
-    final blob = html.Blob([pdf.result!.bytes], 'application/pdf');
-    final url = html.Url.createObjectUrlFromBlob(blob);
-    html.window.open(url, "_blank");
-    html.Url.revokeObjectUrl(url);
-  }
 
   void handleContextMenuAction(
     BuildContext context,
